fix: resolve all 15 serious SwiftLint violations and reduce warnings by 76% [skip ci]
- Extract shared error handling into CloudKitService+ErrorHandling.swift
- Split CloudKitService+Operations.swift (783→4 files) and WriteOperations (419→2 files)
- Split LoggingMiddlewareTests.swift and MistKitClientTests.swift to fix file_length errors
- Rename 7 OpenAPI extension files to match SwiftLint file_name rule
- Fix type_contents_order in QuerySort, QueryFilter, ZoneID, AssetUploadResponse, FieldValue
- Fix conditional_returns_on_newline in all Operations.*.Output.swift files
- Add explicit_acl to test files and Operations Output files
- Add documentation comments to UserIdentity, UserIdentityLookupInfo, NameComponents
- Fix force_try in CloudKitServiceUploadTests+Helpers
- Fix identifier_name (i→requestIndex) in LoggingMiddlewareTests
- Exclude Package.swift from file_name rule in .swiftlint.yml

Result: 322 violations/15 serious → 77 violations/0 serious
