+++ Sources/MistKit/Extensions/OpenAPI/Components.Schemas.FieldValueRequest+MistKit.swift
@@ -1,5 +1,5 @@
 //
-//  Components+FieldValue.swift
+//  Components.Schemas.FieldValueRequest+MistKit.swift
 //  MistKit
 //
 //  Created by Leo Dion.
