+++ Sources/MistKit/PublicTypes/QueryFilter.swift
@@ -32,6 +32,10 @@ import Foundation
 /// Public wrapper for CloudKit query filters
 @available(macOS 11.0, iOS 14.0, tvOS 14.0, watchOS 7.0, *)
 public struct QueryFilter {
+  // MARK: - Internal
+
+  internal let filter: Components.Schemas.Filter
+
   // MARK: - Lifecycle
 
   private init(_ filter: Components.Schemas.Filter) {
@@ -123,7 +127,4 @@ public struct QueryFilter {
     QueryFilter(FilterBuilder.notListMemberBeginsWith(field, prefix))
   }
 
-  // MARK: - Internal
-
-  internal let filter: Components.Schemas.Filter
 }
