+++ Sources/MistKit/Service/CloudKitService+Operations.swift
@@ -7,7 +7,7 @@
 //
 //  Permission is hereby granted, free of charge, to any person
 //  obtaining a copy of this software and associated documentation
-//  files (the “Software”), to deal in the Software without
+//  files (the "Software"), to deal in the Software without
 //  restriction, including without limitation the rights to use,
 //  copy, modify, merge, publish, distribute, sublicense, and/or
 //  sell copies of the Software, and to permit persons to whom the
@@ -17,7 +17,7 @@
 //  The above copyright notice and this permission notice shall be
 //  included in all copies or substantial portions of the Software.
 //
-//  THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND,
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
 //  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
 //  OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
 //  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
@@ -40,193 +40,24 @@ import OpenAPIRuntime
 
 @available(macOS 11.0, iOS 14.0, tvOS 14.0, watchOS 7.0, *)
 extension CloudKitService {
-  /// Fetch current user information
-  public func fetchCurrentUser() async throws(CloudKitError) -> UserInfo {
-    do {
-      let response = try await client.getCurrentUser(
-        .init(
-          path: createGetCurrentUserPath(containerIdentifier: containerIdentifier)
-        )
-      )
-
-      let userData: Components.Schemas.UserResponse =
-        try await responseProcessor.processGetCurrentUserResponse(response)
-      return UserInfo(from: userData)
-    } catch let cloudKitError as CloudKitError {
-      throw cloudKitError
-    } catch let decodingError as DecodingError {
-      MistKitLogger.logError(
-        "JSON decoding failed in fetchCurrentUser: \(decodingError)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.decodingError(decodingError)
-    } catch let urlError as URLError {
-      MistKitLogger.logError(
-        "Network error in fetchCurrentUser: \(urlError)",
-        logger: MistKitLogger.network,
-        shouldRedact: false
-      )
-      throw CloudKitError.networkError(urlError)
-    } catch {
-      MistKitLogger.logError(
-        "Unexpected error in fetchCurrentUser: \(error)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.underlyingError(error)
-    }
-  }
-
-  /// List zones in the user's private database
-  public func listZones() async throws(CloudKitError) -> [ZoneInfo] {
-    do {
-      let response = try await client.listZones(
-        .init(
-          path: createListZonesPath(containerIdentifier: containerIdentifier)
-        )
-      )
-
-      let zonesData: Components.Schemas.ZonesListResponse =
-        try await responseProcessor.processListZonesResponse(response)
-      return zonesData.zones?.compactMap { zone in
-        guard let zoneID = zone.zoneID else {
-          return nil
-        }
-        return ZoneInfo(
-          zoneName: zoneID.zoneName ?? "Unknown",
-          ownerRecordName: zoneID.ownerName,
-          capabilities: []
-        )
-      } ?? []
-    } catch let cloudKitError as CloudKitError {
-      throw cloudKitError
-    } catch let decodingError as DecodingError {
-      MistKitLogger.logError(
-        "JSON decoding failed in listZones: \(decodingError)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.decodingError(decodingError)
-    } catch let urlError as URLError {
-      MistKitLogger.logError(
-        "Network error in listZones: \(urlError)",
-        logger: MistKitLogger.network,
-        shouldRedact: false
-      )
-      throw CloudKitError.networkError(urlError)
-    } catch {
-      MistKitLogger.logError(
-        "Unexpected error in listZones: \(error)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.underlyingError(error)
-    }
-  }
-
-  /// Lookup specific zones by their IDs
-  ///
-  /// Fetches detailed information about multiple zones in a single request.
-  /// Unlike listZones which returns all zones, this operation retrieves
-  /// specific zones identified by their zone IDs.
-  ///
-  /// - Parameter zoneIDs: Array of zone identifiers to lookup
-  /// - Returns: Array of ZoneInfo objects for the requested zones
-  /// - Throws: CloudKitError if the lookup fails
-  ///
-  /// Example:
-  /// ```swift
-  /// let zones = try await service.lookupZones(
-  ///   zoneIDs: [
-  ///     ZoneID(zoneName: "Articles", ownerName: nil),
-  ///     ZoneID(zoneName: "Images", ownerName: nil)
-  ///   ]
-  /// )
-  /// ```
-  public func lookupZones(zoneIDs: [ZoneID]) async throws(CloudKitError) -> [ZoneInfo] {
-    // Validation
-    guard !zoneIDs.isEmpty else {
-      throw CloudKitError.httpErrorWithRawResponse(
-        statusCode: 400,
-        rawResponse: "zoneIDs cannot be empty"
-      )
-    }
-    guard zoneIDs.allSatisfy({ !$0.zoneName.isEmpty }) else {
-      throw CloudKitError.httpErrorWithRawResponse(
-        statusCode: 400,
-        rawResponse: "zoneIDs contains a zone with an empty zoneName"
-      )
-    }
-
-    do {
-      let response = try await client.lookupZones(
-        .init(
-          path: createLookupZonesPath(containerIdentifier: containerIdentifier),
-          body: .json(
-            .init(
-              zones: zoneIDs.map { Components.Schemas.ZoneID(from: $0) }
-            )
-          )
-        )
-      )
-
-      let zonesData: Components.Schemas.ZonesLookupResponse =
-        try await responseProcessor.processLookupZonesResponse(response)
-
-      return zonesData.zones?.compactMap { zone in
-        guard let zoneID = zone.zoneID else {
-          return nil
-        }
-        return ZoneInfo(
-          zoneName: zoneID.zoneName ?? "Unknown",
-          ownerRecordName: zoneID.ownerName,
-          capabilities: []
-        )
-      } ?? []
-    } catch let cloudKitError as CloudKitError {
-      throw cloudKitError
-    } catch let decodingError as DecodingError {
-      MistKitLogger.logError(
-        "JSON decoding failed in lookupZones: \(decodingError)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.decodingError(decodingError)
-    } catch let urlError as URLError {
-      MistKitLogger.logError(
-        "Network error in lookupZones: \(urlError)",
-        logger: MistKitLogger.network,
-        shouldRedact: false
-      )
-      throw CloudKitError.networkError(urlError)
-    } catch {
-      MistKitLogger.logError(
-        "Unexpected error in lookupZones: \(error)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.underlyingError(error)
-    }
-  }
-
   /// Query records from the default zone
   ///
-  /// Queries CloudKit records with optional filtering and sorting. Supports all CloudKit
-  /// filter operations (equals, comparisons, string matching, list operations) and field-based sorting.
+  /// Queries CloudKit records with optional filtering and sorting.
+  /// Supports all CloudKit filter operations (equals, comparisons,
+  /// string matching, list operations) and field-based sorting.
   ///
   /// - Parameters:
   ///   - recordType: The type of records to query (must not be empty)
   ///   - filters: Optional array of filters to apply to the query
   ///   - sortBy: Optional array of sort descriptors
-  ///   - limit: Maximum number of records to return (1-200, defaults to `defaultQueryLimit`)
+  ///   - limit: Maximum number of records to return
+  ///     (1-200, defaults to `defaultQueryLimit`)
   ///   - desiredKeys: Optional array of field names to fetch
   /// - Returns: Array of matching records
   /// - Throws: CloudKitError if validation fails or the request fails
   ///
   /// # Example: Basic Query
   /// ```swift
-  /// // Query all articles
   /// let articles = try await service.queryRecords(
   ///   recordType: "Article"
   /// )
@@ -234,14 +65,11 @@ extension CloudKitService {
   ///
   /// # Example: Query with Filters
   /// ```swift
-  /// // Query published articles from the last week
-  /// let oneWeekAgo = Date().addingTimeInterval(-7 * 24 * 60 * 60)
   /// let recentArticles = try await service.queryRecords(
   ///   recordType: "Article",
   ///   filters: [
   ///     .greaterThan("publishedDate", .date(oneWeekAgo)),
-  ///     .equals("status", .string("published")),
-  ///     .equals("language", .string("en"))
+  ///     .equals("status", .string("published"))
   ///   ],
   ///   limit: 50
   /// )
@@ -249,7 +77,6 @@ extension CloudKitService {
   ///
   /// # Example: Query with Sorting
   /// ```swift
-  /// // Query articles sorted by date (newest first)
   /// let sortedArticles = try await service.queryRecords(
   ///   recordType: "Article",
   ///   sortBy: [.descending("publishedDate")],
@@ -257,41 +84,7 @@ extension CloudKitService {
   /// )
   /// ```
   ///
-  /// # Example: Complex Query with String Matching
-  /// ```swift
-  /// // Search for articles with titles containing "Swift"
-  /// let swiftArticles = try await service.queryRecords(
-  ///   recordType: "Article",
-  ///   filters: [
-  ///     .contains("title", .string("Swift")),
-  ///     .notEquals("status", .string("draft"))
-  ///   ],
-  ///   sortBy: [.descending("publishedDate")],
-  ///   desiredKeys: ["title", "publishedDate", "author"]
-  /// )
-  /// ```
-  ///
-  /// # Example: List Operations
-  /// ```swift
-  /// // Query articles with specific tags
-  /// let taggedArticles = try await service.queryRecords(
-  ///   recordType: "Article",
-  ///   filters: [
-  ///     .in("category", [.string("Technology"), .string("Programming")]),
-  ///     .greaterThanOrEquals("viewCount", .int64(1000))
-  ///   ]
-  /// )
-  /// ```
-  ///
-  /// # Available Filter Operations
-  /// - Equality: `.equals()`, `.notEquals()`
-  /// - Comparison: `.lessThan()`, `.lessThanOrEquals()`, `.greaterThan()`, `.greaterThanOrEquals()`
-  /// - String: `.beginsWith()`, `.contains()`, `.endsWith()`
-  /// - List: `.in()`, `.notIn()`
-  /// - Negation: `.not()`
-  ///
-  /// - Note: For large result sets, consider using pagination (see GitHub issue #145)
-  /// - Note: To query custom zones, see GitHub issue #146
+  /// - Note: For large result sets, consider using pagination
   public func queryRecords(
     recordType: String,
     filters: [QueryFilter]? = nil,
@@ -299,10 +92,8 @@ extension CloudKitService {
     limit: Int? = nil,
     desiredKeys: [String]? = nil
   ) async throws(CloudKitError) -> [RecordInfo] {
-    // Use provided limit or fall back to default configuration
     let effectiveLimit = limit ?? defaultQueryLimit
 
-    // Validate input parameters
     guard !recordType.isEmpty else {
       throw CloudKitError.httpErrorWithRawResponse(
         statusCode: 400,
@@ -313,17 +104,24 @@ extension CloudKitService {
     guard effectiveLimit > 0 && effectiveLimit <= 200 else {
       throw CloudKitError.httpErrorWithRawResponse(
         statusCode: 400,
-        rawResponse: "limit must be between 1 and 200, got \(effectiveLimit)"
+        rawResponse:
+          "limit must be between 1 and 200, got \(effectiveLimit)"
       )
     }
 
-    let componentsFilters = filters?.map { Components.Schemas.Filter(from: $0) }
-    let componentsSorts = sortBy?.map { Components.Schemas.Sort(from: $0) }
+    let componentsFilters = filters?.map {
+      Components.Schemas.Filter(from: $0)
+    }
+    let componentsSorts = sortBy?.map {
+      Components.Schemas.Sort(from: $0)
+    }
 
     do {
       let response = try await client.queryRecords(
         .init(
-          path: createQueryRecordsPath(containerIdentifier: containerIdentifier),
+          path: createQueryRecordsPath(
+            containerIdentifier: containerIdentifier
+          ),
           body: .json(
             .init(
               zoneID: .init(zoneName: "_defaultZone"),
@@ -342,338 +140,8 @@ extension CloudKitService {
       let recordsData: Components.Schemas.QueryResponse =
         try await responseProcessor.processQueryRecordsResponse(response)
       return recordsData.records?.compactMap { RecordInfo(from: $0) } ?? []
-    } catch let cloudKitError as CloudKitError {
-      throw cloudKitError
-    } catch let decodingError as DecodingError {
-      // Log detailed decoding error information
-      MistKitLogger.logError(
-        "JSON decoding failed in queryRecords: \(decodingError)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-
-      // Log detailed context based on error type
-      switch decodingError {
-      case .keyNotFound(let key, let context):
-        MistKitLogger.logDebug(
-          "Missing key: \(key), Context: \(context.debugDescription), Coding path: \(context.codingPath)",
-          logger: MistKitLogger.api,
-          shouldRedact: false
-        )
-      case .typeMismatch(let type, let context):
-        MistKitLogger.logDebug(
-          "Type mismatch: expected \(type), Context: \(context.debugDescription), Coding path: \(context.codingPath)",
-          logger: MistKitLogger.api,
-          shouldRedact: false
-        )
-      case .valueNotFound(let type, let context):
-        MistKitLogger.logDebug(
-          "Value not found: expected \(type), Context: \(context.debugDescription), Coding path: \(context.codingPath)",
-          logger: MistKitLogger.api,
-          shouldRedact: false
-        )
-      case .dataCorrupted(let context):
-        MistKitLogger.logDebug(
-          "Data corrupted, Context: \(context.debugDescription), Coding path: \(context.codingPath)",
-          logger: MistKitLogger.api,
-          shouldRedact: false
-        )
-      @unknown default:
-        MistKitLogger.logDebug(
-          "Unknown decoding error type",
-          logger: MistKitLogger.api,
-          shouldRedact: false
-        )
-      }
-
-      throw CloudKitError.decodingError(decodingError)
-    } catch let urlError as URLError {
-      // Log network error information
-      MistKitLogger.logError(
-        "Network error in queryRecords: \(urlError)",
-        logger: MistKitLogger.network,
-        shouldRedact: false
-      )
-
-      throw CloudKitError.networkError(urlError)
     } catch {
-      // Log unexpected errors
-      MistKitLogger.logError(
-        "Unexpected error in queryRecords: \(error)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-
-      // Log additional debugging details
-      MistKitLogger.logDebug(
-        "Error type: \(type(of: error)), Description: \(String(reflecting: error))",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-
-      throw CloudKitError.underlyingError(error)
-    }
-  }
-
-  /// Fetch zone changes since a sync token
-  ///
-  /// Retrieves all zones that have changed since the provided sync token.
-  /// Use this for efficient incremental sync at the zone level.
-  ///
-  /// - Parameter syncToken: Optional token from previous fetch (nil = initial fetch)
-  /// - Returns: ZoneChangesResult containing changed zones and new sync token
-  /// - Throws: CloudKitError if the fetch fails
-  ///
-  /// Example - Initial Sync:
-  /// ```swift
-  /// let result = try await service.fetchZoneChanges()
-  /// // Store result.syncToken for next fetch
-  /// processZones(result.zones)
-  /// ```
-  ///
-  /// Example - Incremental Sync:
-  /// ```swift
-  /// let result = try await service.fetchZoneChanges(
-  ///   syncToken: previousToken
-  /// )
-  /// ```
-  public func fetchZoneChanges(
-    syncToken: String? = nil
-  ) async throws(CloudKitError) -> ZoneChangesResult {
-    do {
-      let response = try await client.fetchZoneChanges(
-        .init(
-          path: createFetchZoneChangesPath(containerIdentifier: containerIdentifier),
-          body: .json(
-            .init(
-              syncToken: syncToken
-            )
-          )
-        )
-      )
-
-      let changesData: Components.Schemas.ZoneChangesResponse =
-        try await responseProcessor.processFetchZoneChangesResponse(response)
-
-      return ZoneChangesResult(from: changesData)
-    } catch let cloudKitError as CloudKitError {
-      throw cloudKitError
-    } catch let decodingError as DecodingError {
-      MistKitLogger.logError(
-        "JSON decoding failed in fetchZoneChanges: \(decodingError)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.decodingError(decodingError)
-    } catch let urlError as URLError {
-      MistKitLogger.logError(
-        "Network error in fetchZoneChanges: \(urlError)",
-        logger: MistKitLogger.network,
-        shouldRedact: false
-      )
-      throw CloudKitError.networkError(urlError)
-    } catch {
-      MistKitLogger.logError(
-        "Unexpected error in fetchZoneChanges: \(error)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.underlyingError(error)
-    }
-  }
-
-  /// Fetch record changes since a sync token
-  ///
-  /// Retrieves all records that have changed (created, updated, or deleted)
-  /// since the provided sync token. Use this for efficient incremental sync
-  /// operations rather than repeatedly querying all records.
-  ///
-  /// - Parameters:
-  ///   - zoneID: Optional zone to fetch changes from (defaults to _defaultZone)
-  ///   - syncToken: Optional token from previous fetch (nil = initial fetch)
-  ///   - resultsLimit: Optional maximum number of records (1-200)
-  /// - Returns: RecordChangesResult containing changed records and new sync token
-  /// - Throws: CloudKitError if the fetch fails
-  ///
-  /// Example - Initial Sync:
-  /// ```swift
-  /// let result = try await service.fetchRecordChanges()
-  /// // Store result.syncToken for next fetch
-  /// processRecords(result.records)
-  /// ```
-  ///
-  /// Example - Incremental Sync:
-  /// ```swift
-  /// let result = try await service.fetchRecordChanges(
-  ///   syncToken: previousToken
-  /// )
-  /// if result.moreComing {
-  ///   // More changes available, fetch again with new token
-  ///   let next = try await service.fetchRecordChanges(
-  ///     syncToken: result.syncToken
-  ///   )
-  /// }
-  /// ```
-  ///
-  /// - Note: If moreComing is true, call again with the returned syncToken
-  ///         to fetch remaining changes
-  public func fetchRecordChanges(
-    zoneID: ZoneID? = nil,
-    syncToken: String? = nil,
-    resultsLimit: Int? = nil
-  ) async throws(CloudKitError) -> RecordChangesResult {
-    // Validate limit if provided
-    if let limit = resultsLimit {
-      guard limit > 0 && limit <= 200 else {
-        throw CloudKitError.httpErrorWithRawResponse(
-          statusCode: 400,
-          rawResponse: "resultsLimit must be between 1 and 200, got \(limit)"
-        )
-      }
-    }
-
-    // Use provided zoneID or default zone
-    let effectiveZoneID = zoneID ?? .defaultZone
-
-    do {
-      let response = try await client.fetchRecordChanges(
-        .init(
-          path: createFetchRecordChangesPath(containerIdentifier: containerIdentifier),
-          body: .json(
-            .init(
-              zoneID: Components.Schemas.ZoneID(from: effectiveZoneID),
-              syncToken: syncToken,
-              resultsLimit: resultsLimit
-            )
-          )
-        )
-      )
-
-      let changesData: Components.Schemas.ChangesResponse =
-        try await responseProcessor.processFetchRecordChangesResponse(response)
-
-      return RecordChangesResult(from: changesData)
-    } catch let cloudKitError as CloudKitError {
-      throw cloudKitError
-    } catch let decodingError as DecodingError {
-      MistKitLogger.logError(
-        "JSON decoding failed in fetchRecordChanges: \(decodingError)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.decodingError(decodingError)
-    } catch let urlError as URLError {
-      MistKitLogger.logError(
-        "Network error in fetchRecordChanges: \(urlError)",
-        logger: MistKitLogger.network,
-        shouldRedact: false
-      )
-      throw CloudKitError.networkError(urlError)
-    } catch {
-      MistKitLogger.logError(
-        "Unexpected error in fetchRecordChanges: \(error)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.underlyingError(error)
-    }
-  }
-
-  /// Fetch all record changes, handling pagination automatically
-  ///
-  /// Convenience method that automatically fetches all available changes
-  /// by following the moreComing flag and making multiple requests if needed.
-  ///
-  /// - Parameters:
-  ///   - zoneID: Optional zone to fetch changes from (defaults to _defaultZone)
-  ///   - syncToken: Optional token from previous fetch (nil = initial fetch)
-  ///   - resultsLimit: Optional maximum records per request (1-200)
-  /// - Returns: Array of all changed records and final sync token
-  /// - Throws: CloudKitError if any fetch fails
-  ///
-  /// Example:
-  /// ```swift
-  /// let (records, newToken) = try await service.fetchAllRecordChanges(
-  ///   syncToken: lastSyncToken
-  /// )
-  /// // Process all records
-  /// processRecords(records)
-  /// // Store newToken for next sync
-  /// ```
-  ///
-  /// - Warning: For zones with many changes, this may make multiple requests
-  ///           and return a large array. Consider using fetchRecordChanges()
-  ///           with manual pagination for better memory control.
-  /// - Note: Makes sequential requests with no backoff or cooperative cancellation
-  ///         between pages. For fine-grained control, use `fetchRecordChanges(syncToken:)` directly.
-  public func fetchAllRecordChanges(
-    zoneID: ZoneID? = nil,
-    syncToken: String? = nil,
-    resultsLimit: Int? = nil
-  ) async throws(CloudKitError) -> (records: [RecordInfo], syncToken: String?) {
-    var allRecords: [RecordInfo] = []
-    var currentToken = syncToken
-    var moreComing = true
-
-    while moreComing {
-      let result = try await fetchRecordChanges(
-        zoneID: zoneID,
-        syncToken: currentToken,
-        resultsLimit: resultsLimit
-      )
-
-      allRecords.append(contentsOf: result.records)
-      currentToken = result.syncToken
-      moreComing = result.moreComing
-    }
-
-    return (allRecords, currentToken)
-  }
-
-  /// Discover user identities by email addresses or record names
-  public func discoverUserIdentities(
-    lookupInfos: [UserIdentityLookupInfo]
-  ) async throws(CloudKitError) -> [UserIdentity] {
-    do {
-      let response = try await client.discoverUserIdentities(
-        .init(
-          path: createDiscoverUserIdentitiesPath(containerIdentifier: containerIdentifier),
-          body: .json(
-            .init(
-              lookupInfos: lookupInfos.map {
-                .init(emailAddress: $0.emailAddress, phoneNumber: $0.phoneNumber, userRecordName: $0.userRecordName)
-              }
-            )
-          )
-        )
-      )
-
-      let discoverData: Components.Schemas.DiscoverResponse =
-        try await responseProcessor.processDiscoverUserIdentitiesResponse(response)
-      return discoverData.users?.map(UserIdentity.init(from:)) ?? []
-    } catch let cloudKitError as CloudKitError {
-      throw cloudKitError
-    } catch let decodingError as DecodingError {
-      MistKitLogger.logError(
-        "JSON decoding failed in discoverUserIdentities: \(decodingError)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.decodingError(decodingError)
-    } catch let urlError as URLError {
-      MistKitLogger.logError(
-        "Network error in discoverUserIdentities: \(urlError)",
-        logger: MistKitLogger.network,
-        shouldRedact: false
-      )
-      throw CloudKitError.networkError(urlError)
-    } catch {
-      MistKitLogger.logError(
-        "Unexpected error in discoverUserIdentities: \(error)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.underlyingError(error)
+      throw mapToCloudKitError(error, context: "queryRecords")
     }
   }
 
@@ -689,7 +157,9 @@ extension CloudKitService {
     do {
       let response = try await client.modifyRecords(
         .init(
-          path: createModifyRecordsPath(containerIdentifier: containerIdentifier),
+          path: createModifyRecordsPath(
+            containerIdentifier: containerIdentifier
+          ),
           body: .json(
             .init(
               operations: operations,
@@ -702,29 +172,8 @@ extension CloudKitService {
       let modifyData: Components.Schemas.ModifyResponse =
         try await responseProcessor.processModifyRecordsResponse(response)
       return modifyData.records?.compactMap { RecordInfo(from: $0) } ?? []
-    } catch let cloudKitError as CloudKitError {
-      throw cloudKitError
-    } catch let decodingError as DecodingError {
-      MistKitLogger.logError(
-        "JSON decoding failed in modifyRecords: \(decodingError)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.decodingError(decodingError)
-    } catch let urlError as URLError {
-      MistKitLogger.logError(
-        "Network error in modifyRecords: \(urlError)",
-        logger: MistKitLogger.network,
-        shouldRedact: false
-      )
-      throw CloudKitError.networkError(urlError)
     } catch {
-      MistKitLogger.logError(
-        "Unexpected error in modifyRecords: \(error)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.underlyingError(error)
+      throw mapToCloudKitError(error, context: "modifyRecords")
     }
   }
 
@@ -736,7 +185,9 @@ extension CloudKitService {
     do {
       let response = try await client.lookupRecords(
         .init(
-          path: createLookupRecordsPath(containerIdentifier: containerIdentifier),
+          path: createLookupRecordsPath(
+            containerIdentifier: containerIdentifier
+          ),
           body: .json(
             .init(
               records: recordNames.map { recordName in
@@ -753,29 +204,8 @@ extension CloudKitService {
       let lookupData: Components.Schemas.LookupResponse =
         try await responseProcessor.processLookupRecordsResponse(response)
       return lookupData.records?.compactMap { RecordInfo(from: $0) } ?? []
-    } catch let cloudKitError as CloudKitError {
-      throw cloudKitError
-    } catch let decodingError as DecodingError {
-      MistKitLogger.logError(
-        "JSON decoding failed in lookupRecords: \(decodingError)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.decodingError(decodingError)
-    } catch let urlError as URLError {
-      MistKitLogger.logError(
-        "Network error in lookupRecords: \(urlError)",
-        logger: MistKitLogger.network,
-        shouldRedact: false
-      )
-      throw CloudKitError.networkError(urlError)
     } catch {
-      MistKitLogger.logError(
-        "Unexpected error in lookupRecords: \(error)",
-        logger: MistKitLogger.api,
-        shouldRedact: false
-      )
-      throw CloudKitError.underlyingError(error)
+      throw mapToCloudKitError(error, context: "lookupRecords")
     }
   }
 }
