+++ Sources/MistKit/Service/NameComponents.swift
@@ -7,7 +7,7 @@
 //
 //  Permission is hereby granted, free of charge, to any person
 //  obtaining a copy of this software and associated documentation
-//  files (the "Software"), to deal in the Software without
+//  files (the “Software”), to deal in the Software without
 //  restriction, including without limitation the rights to use,
 //  copy, modify, merge, publish, distribute, sublicense, and/or
 //  sell copies of the Software, and to permit persons to whom the
@@ -17,7 +17,7 @@
 //  The above copyright notice and this permission notice shall be
 //  included in all copies or substantial portions of the Software.
 //
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+//  THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND,
 //  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
 //  OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
 //  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
@@ -29,12 +29,19 @@
 
 /// The parts of a user's name from CloudKit user discovery.
 public struct NameComponents: Codable, Sendable {
+  /// The name prefix (e.g., "Dr.", "Mr.")
   public let namePrefix: String?
+  /// The given (first) name
   public let givenName: String?
+  /// The middle name
   public let middleName: String?
+  /// The family (last) name
   public let familyName: String?
+  /// The name suffix (e.g., "Jr.", "III")
   public let nameSuffix: String?
+  /// The nickname
   public let nickname: String?
+  /// The phonetic representation of the name
   public let phoneticRepresentation: String?
 
   internal init(from schema: Components.Schemas.NameComponents) {
@@ -47,6 +54,15 @@ public struct NameComponents: Codable, Sendable {
     self.phoneticRepresentation = schema.phoneticRepresentation
   }
 
+  /// Initialize name components
+  /// - Parameters:
+  ///   - namePrefix: The name prefix
+  ///   - givenName: The given name
+  ///   - middleName: The middle name
+  ///   - familyName: The family name
+  ///   - nameSuffix: The name suffix
+  ///   - nickname: The nickname
+  ///   - phoneticRepresentation: The phonetic representation
   public init(
     namePrefix: String? = nil,
     givenName: String? = nil,
