+++ Sources/MistKit/Service/Operations.fetchRecordChanges.Output.swift
@@ -28,51 +28,95 @@
 //
 
 extension Operations.fetchRecordChanges.Output: CloudKitResponseType {
-  var badRequestResponse: Components.Responses.BadRequest? {
-    if case .badRequest(let response) = self { return response } else { return nil }
+  internal var badRequestResponse: Components.Responses.BadRequest? {
+    if case .badRequest(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
-  var unauthorizedResponse: Components.Responses.Unauthorized? {
-    if case .unauthorized(let response) = self { return response } else { return nil }
+  internal var unauthorizedResponse: Components.Responses.Unauthorized? {
+    if case .unauthorized(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
-  var forbiddenResponse: Components.Responses.Forbidden? {
-    if case .forbidden(let response) = self { return response } else { return nil }
+  internal var forbiddenResponse: Components.Responses.Forbidden? {
+    if case .forbidden(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
-  var notFoundResponse: Components.Responses.NotFound? {
-    if case .notFound(let response) = self { return response } else { return nil }
+  internal var notFoundResponse: Components.Responses.NotFound? {
+    if case .notFound(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
-  var conflictResponse: Components.Responses.Conflict? {
-    if case .conflict(let response) = self { return response } else { return nil }
+  internal var conflictResponse: Components.Responses.Conflict? {
+    if case .conflict(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
-  var preconditionFailedResponse: Components.Responses.PreconditionFailed? {
-    if case .preconditionFailed(let response) = self { return response } else { return nil }
+  internal var preconditionFailedResponse: Components.Responses.PreconditionFailed? {
+    if case .preconditionFailed(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
-  var contentTooLargeResponse: Components.Responses.RequestEntityTooLarge? {
-    if case .contentTooLarge(let response) = self { return response } else { return nil }
+  internal var contentTooLargeResponse: Components.Responses.RequestEntityTooLarge? {
+    if case .contentTooLarge(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
-  var misdirectedRequestResponse: Components.Responses.UnprocessableEntity? {
-    if case .misdirectedRequest(let response) = self { return response } else { return nil }
+  internal var misdirectedRequestResponse: Components.Responses.UnprocessableEntity? {
+    if case .misdirectedRequest(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
-  var tooManyRequestsResponse: Components.Responses.TooManyRequests? {
-    if case .tooManyRequests(let response) = self { return response } else { return nil }
+  internal var tooManyRequestsResponse: Components.Responses.TooManyRequests? {
+    if case .tooManyRequests(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
-  var isOk: Bool {
-    if case .ok = self { return true } else { return false }
+  internal var isOk: Bool {
+    if case .ok = self {
+      return true
+    } else {
+      return false
+    }
   }
 
-  var undocumentedStatusCode: Int? {
-    if case .undocumented(let statusCode, _) = self { return statusCode } else { return nil }
+  internal var undocumentedStatusCode: Int? {
+    if case .undocumented(let statusCode, _) = self {
+      return statusCode
+    } else {
+      return nil
+    }
   }
 
   // fetchRecordChanges has most error responses except 500/503
-  var internalServerErrorResponse: Components.Responses.InternalServerError? { nil }
-  var serviceUnavailableResponse: Components.Responses.ServiceUnavailable? { nil }
+  internal var internalServerErrorResponse: Components.Responses.InternalServerError? { nil }
+  internal var serviceUnavailableResponse: Components.Responses.ServiceUnavailable? { nil }
 }
