+++ Sources/MistKit/Service/Operations.fetchZoneChanges.Output.swift
@@ -7,7 +7,7 @@
 //
 //  Permission is hereby granted, free of charge, to any person
 //  obtaining a copy of this software and associated documentation
-//  files (the "Software"), to deal in the Software without
+//  files (the “Software”), to deal in the Software without
 //  restriction, including without limitation the rights to use,
 //  copy, modify, merge, publish, distribute, sublicense, and/or
 //  sell copies of the Software, and to permit persons to whom the
@@ -17,7 +17,7 @@
 //  The above copyright notice and this permission notice shall be
 //  included in all copies or substantial portions of the Software.
 //
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+//  THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND,
 //  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
 //  OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
 //  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
@@ -29,19 +29,35 @@
 
 extension Operations.fetchZoneChanges.Output: CloudKitResponseType {
   internal var badRequestResponse: Components.Responses.BadRequest? {
-    if case .badRequest(let response) = self { return response } else { return nil }
+    if case .badRequest(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
   internal var unauthorizedResponse: Components.Responses.Unauthorized? {
-    if case .unauthorized(let response) = self { return response } else { return nil }
+    if case .unauthorized(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
   internal var isOk: Bool {
-    if case .ok = self { return true } else { return false }
+    if case .ok = self {
+      return true
+    } else {
+      return false
+    }
   }
 
   internal var undocumentedStatusCode: Int? {
-    if case .undocumented(let statusCode, _) = self { return statusCode } else { return nil }
+    if case .undocumented(let statusCode, _) = self {
+      return statusCode
+    } else {
+      return nil
+    }
   }
 
   // fetchZoneChanges only defines 200, 400, 401 responses
