+++ Sources/MistKit/Service/Operations.listZones.Output.swift
@@ -29,54 +29,106 @@
 
 extension Operations.listZones.Output: CloudKitResponseType {
   internal var badRequestResponse: Components.Responses.BadRequest? {
-    if case .badRequest(let response) = self { return response } else { return nil }
+    if case .badRequest(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
   internal var unauthorizedResponse: Components.Responses.Unauthorized? {
-    if case .unauthorized(let response) = self { return response } else { return nil }
+    if case .unauthorized(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
   internal var forbiddenResponse: Components.Responses.Forbidden? {
-    if case .forbidden(let response) = self { return response } else { return nil }
+    if case .forbidden(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
   internal var notFoundResponse: Components.Responses.NotFound? {
-    if case .notFound(let response) = self { return response } else { return nil }
+    if case .notFound(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
   internal var conflictResponse: Components.Responses.Conflict? {
-    if case .conflict(let response) = self { return response } else { return nil }
+    if case .conflict(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
   internal var preconditionFailedResponse: Components.Responses.PreconditionFailed? {
-    if case .preconditionFailed(let response) = self { return response } else { return nil }
+    if case .preconditionFailed(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
   internal var contentTooLargeResponse: Components.Responses.RequestEntityTooLarge? {
-    if case .contentTooLarge(let response) = self { return response } else { return nil }
+    if case .contentTooLarge(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
   internal var misdirectedRequestResponse: Components.Responses.UnprocessableEntity? {
-    if case .misdirectedRequest(let response) = self { return response } else { return nil }
+    if case .misdirectedRequest(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
   internal var tooManyRequestsResponse: Components.Responses.TooManyRequests? {
-    if case .tooManyRequests(let response) = self { return response } else { return nil }
+    if case .tooManyRequests(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
   internal var internalServerErrorResponse: Components.Responses.InternalServerError? {
-    if case .internalServerError(let response) = self { return response } else { return nil }
+    if case .internalServerError(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
   internal var serviceUnavailableResponse: Components.Responses.ServiceUnavailable? {
-    if case .serviceUnavailable(let response) = self { return response } else { return nil }
+    if case .serviceUnavailable(let response) = self {
+      return response
+    } else {
+      return nil
+    }
   }
 
   internal var isOk: Bool {
-    if case .ok = self { return true } else { return false }
+    if case .ok = self {
+      return true
+    } else {
+      return false
+    }
   }
 
   internal var undocumentedStatusCode: Int? {
-    if case .undocumented(let statusCode, _) = self { return statusCode } else { return nil }
+    if case .undocumented(let statusCode, _) = self {
+      return statusCode
+    } else {
+      return nil
+    }
   }
 }
