+++ Sources/MistKit/Service/ZoneChangesResult.swift
@@ -7,7 +7,7 @@
 //
 //  Permission is hereby granted, free of charge, to any person
 //  obtaining a copy of this software and associated documentation
-//  files (the "Software"), to deal in the Software without
+//  files (the “Software”), to deal in the Software without
 //  restriction, including without limitation the rights to use,
 //  copy, modify, merge, publish, distribute, sublicense, and/or
 //  sell copies of the Software, and to permit persons to whom the
@@ -17,7 +17,7 @@
 //  The above copyright notice and this permission notice shall be
 //  included in all copies or substantial portions of the Software.
 //
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+//  THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND,
 //  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
 //  OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
 //  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
@@ -47,16 +47,17 @@ public struct ZoneChangesResult: Sendable {
   }
 
   internal init(from response: Components.Schemas.ZoneChangesResponse) {
-    self.zones = response.zones?.compactMap { zonePayload in
-      guard let zoneID = zonePayload.zoneID else {
-        return nil
-      }
-      return ZoneInfo(
-        zoneName: zoneID.zoneName ?? "Unknown",
-        ownerRecordName: zoneID.ownerName,
-        capabilities: []  // CloudKit Web Services zone-changes responses omit capabilities
-      )
-    } ?? []
+    self.zones =
+      response.zones?.compactMap { zonePayload in
+        guard let zoneID = zonePayload.zoneID else {
+          return nil
+        }
+        return ZoneInfo(
+          zoneName: zoneID.zoneName ?? "Unknown",
+          ownerRecordName: zoneID.ownerName,
+          capabilities: []  // CloudKit Web Services zone-changes responses omit capabilities
+        )
+      } ?? []
     self.syncToken = response.syncToken
   }
 }
