+++ Tests/MistKitTests/Service/CloudKitServiceFetchZoneChangesTests+Helpers.swift
@@ -92,7 +92,7 @@ extension ResponseConfig {
       zones.append([
         "zoneID": [
           "zoneName": "test-zone-\(index)",
-          "ownerName": "_defaultOwner"
+          "ownerName": "_defaultOwner",
         ]
       ])
     }
